added to pages, created additional components for pages, and modified app.js
