+++ client/src/components/Login/index.js
@@ -1,0 +1,19 @@
+import React from 'react';
+
+export default function Login() {
+  return(
+    <form>
+      <label>
+        <p>Username</p>
+        <input type="text" />
+      </label>
+      <label>
+        <p>Password</p>
+        <input type="password" />
+      </label>
+      <div>
+        <button type="submit">Submit</button>
+      </div>
+    </form>
+  )
+}
