+++ client/src/components/Wrapper/style.css
@@ -1,0 +1,4 @@
+.wrapper {
+  min-height: 100%;
+  padding-bottom: 100px;
+}
