+++ client/src/pages/About.js
@@ -1,0 +1,20 @@
+import React from "react";
+import Login from "../components/Login";
+import {Col, Row, Container } from "../components/Grid";
+import Jumbotron from "../components/Jumbotron";
+
+function About () {
+    return (
+        <div>
+            <Container fluid>
+                <Jumbotron /> 
+                <Row>
+                    <Col>
+                    </Col>
+                </Row>
+            </Container>
+        </div>
+    )
+}
+
+export default About;
